add the Tour service & repository

+++ src/main/java/com/logistique/gestiontournees/repository/TourRepository.java
@@ -3,5 +3,5 @@ package com.logistique.gestiontournees.repository;
 import com.logistique.gestiontournees.entity.Tour;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TourRepository extends JpaRepository<Tour, Integer> {
+public interface TourRepository extends JpaRepository<Tour, Long> {
 }
